pyproject.toml
[build-system]
requires      = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name            = "dxf-pipeline"
version         = "0.1.0"
requires-python = ">=3.10"
dependencies    = [
    "ezdxf>=1.3",
    "pillow>=10.0",
    "cairosvg>=2.7",
]

[tool.pytest.ini_options]
pythonpath = ["pipeline"]
testpaths  = ["tests"]
addopts    = "--cov=pipeline --cov-report=term-missing --cov-fail-under=100"
markers    = [
    "integration: requires cairosvg and does real SVG rendering",
]

[tool.coverage.run]
source = ["pipeline"]

[tool.coverage.report]
fail_under = 100
exclude_lines = [
    "pragma: no cover",
    'if __name__ == "__main__":',
]

[tool.mypy]
python_version         = "3.10"
ignore_missing_imports = true
check_untyped_defs     = true
warn_unused_ignores    = true

[tool.ruff]
line-length    = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
